+++ tasks/upload_tasks.py
@@ -1,4 +1,5 @@
 # tasks/upload_tasks.py
+
 """claude async control"""
 
 # ===== STANDARD LIBRARY IMPORTS =====
@@ -78,19 +79,22 @@ from utils.document_loaders.performance import create_optimized_processor, Batch
 from utils.metrics import MetricsCollector, Timer
 from utils.connection_pool import ConnectionPoolManager
 from utils.memory_manager import MemoryManager # Kept for health checks
-from tasks.celery_app import (
-    run_async_in_worker,
-    get_global_async_db_pool,
-    get_global_redis_pool,
-    init_async_pools,
-    get_db_connection,      # ← Context manager
-    get_redis_connection    # ← Context manager
-)
-# Import health checks from the shared module:
-from tasks.pool_utils import (
-    check_async_db_pool_health,
-    check_redis_pool_health
-)
+
+from tasks.celery_app import run_async_in_worker
+from tasks.database import get_db_connection, get_redis_connection, get_global_async_db_pool, get_global_redis_pool, init_async_pools, check_db_pool_health, check_redis_pool_health
+
+# from tasks.celery_app import (
+#     run_async_in_worker,
+#     get_global_async_db_pool,
+#     get_global_redis_pool,
+#     init_async_pools,
+#     get_db_connection,      # ← Context manager
+#     get_redis_connection    # ← Context manager
+# )
+# # Import health checks from the shared module:
+# from tasks.pool_utils import (
+#     check_redis_pool_health
+# )
 
 # ——— Logging & Env Load ———————————————————————————————————————————————————————————
 logger = get_task_logger(__name__)
